refactor(aspect): Restructure preprocessing code
Clearly separate the source code into reusable parts and code specific
to a given preprocessing mode.
