+++ dwyu/aspect/private/preprocessing/fast_parsing/BUILD
@@ -1,0 +1,8 @@
+load("@rules_cc//cc:cc_library.bzl", "cc_library")
+
+cc_library(
+    name = "extract_includes",
+    srcs = ["extract_includes.cpp"],
+    hdrs = ["extract_includes.h"],
+    visibility = ["//dwyu/aspect/private/preprocessing:__subpackages__"],
+)
