+++ dwyu/aspect/private/preprocessing/fast_parsing/extract_includes.h
@@ -1,5 +1,5 @@
-#ifndef DWYU_ASPECT_PRIVATE_PREPROCESSING_EXTRACT_INCLUDES_H
-#define DWYU_ASPECT_PRIVATE_PREPROCESSING_EXTRACT_INCLUDES_H
+#ifndef DWYU_ASPECT_PRIVATE_PREPROCESSING_FAST_PARSING_EXTRACT_INCLUDES_H
+#define DWYU_ASPECT_PRIVATE_PREPROCESSING_FAST_PARSING_EXTRACT_INCLUDES_H
 
 #include <istream>
 #include <set>
