+++ dwyu/aspect/private/preprocessing/lib/included_file.h
@@ -1,5 +1,5 @@
-#ifndef DWYU_ASPECT_PRIVATE_PREPROCESSING_INCLUDED_FILE_H
-#define DWYU_ASPECT_PRIVATE_PREPROCESSING_INCLUDED_FILE_H
+#ifndef DWYU_ASPECT_PRIVATE_PREPROCESSING_LIB_INCLUDED_FILE_H
+#define DWYU_ASPECT_PRIVATE_PREPROCESSING_LIB_INCLUDED_FILE_H
 
 #include <boost/filesystem/path.hpp>
